requirements.txt
azure-common==1.1.28
azure-core==1.26.3
azure-cosmos==4.3.1
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-data-tables==12.4.2
azure-functions==1.13.2
azure-nspkg==3.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
idna==3.4
isodate==0.6.1
multidict==6.0.4
pycparser==2.21
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
